LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-history-go-0-expected.txt



FAIL history.go(0) fires the navigate event assert_equals: expected "reload" but got "push"
